snapshot test Parameters::import() and fix typos
