+++ config/tests/config_tests.rs
@@ -4,12 +4,14 @@
 use std::collections::BTreeMap;
 
 use config::{
-    Committee, ConsensusAPIGrpcParameters, Epoch, Parameters, PrimaryAddresses,
+    Committee, ConsensusAPIGrpcParameters, Epoch, Import, Parameters, PrimaryAddresses,
     PrometheusMetricsParameters, Stake,
 };
 use crypto::{ed25519::Ed25519PublicKey, traits::KeyPair};
 use insta::assert_json_snapshot;
 use rand::seq::SliceRandom;
+use std::{fs::File, io::Write};
+use tempfile::tempdir;
 use test_utils::make_authority_with_port_getter;
 
 #[test]
@@ -22,7 +24,7 @@ fn update_primary_network_info_test() {
     for err in res {
         assert!(matches!(
             err,
-            config::ComitteeUpdateError::MissingFromUpdate(_)
+            config::CommitteeUpdateError::MissingFromUpdate(_)
         ))
     }
 
@@ -40,8 +42,8 @@ fn update_primary_network_info_test() {
         // we'll get the two collections reporting missing from each other
         assert!(matches!(
             err,
-            config::ComitteeUpdateError::NotInCommittee(_)
-                | config::ComitteeUpdateError::MissingFromUpdate(_)
+            config::CommitteeUpdateError::NotInCommittee(_)
+                | config::CommitteeUpdateError::MissingFromUpdate(_)
         ))
     }
 
@@ -59,7 +61,7 @@ fn update_primary_network_info_test() {
     for err in res2 {
         assert!(matches!(
             err,
-            config::ComitteeUpdateError::DifferentStake(_)
+            config::CommitteeUpdateError::DifferentStake(_)
         ))
     }
 
@@ -139,4 +141,48 @@ fn commmittee_snapshot_matches() {
     let mut settings = insta::Settings::clone_current();
     settings.set_sort_maps(true);
     settings.bind(|| assert_json_snapshot!("committee", committee));
+}
+
+#[test]
+fn parameters_import_snapshot_matches() {
+    // GIVEN
+    let input = r#"{
+         "header_size": 1000,
+         "max_header_delay": "100ms",
+         "gc_depth": 50,
+         "sync_retry_delay": "5s",
+         "sync_retry_nodes": 3,
+         "batch_size": 500000,
+         "max_batch_delay": "100ms",
+         "block_synchronizer": {
+             "certificates_synchronize_timeout": "2s",
+             "payload_synchronize_timeout": "3_000ms",
+             "payload_availability_timeout": "4_000ms",
+             "handler_certificate_deliver_timeout": "1_000ms"
+         },
+         "consensus_api_grpc": {
+             "socket_addr": "/ip4/127.0.0.1/tcp/0/http",
+             "get_collections_timeout": "5_000ms",
+             "remove_collections_timeout": "5_000ms"
+         },
+         "max_concurrent_requests": 500000,
+         "prometheus_metrics": {
+             "socket_addr": "127.0.0.1:0"
+         }
+      }"#;
+
+    // AND temporary file
+    let dir = tempdir().expect("Couldn't create tempdir");
+
+    let file_path = dir.path().join("temp-properties.json");
+    let mut file = File::create(file_path.clone()).expect("Couldn't create temp file");
+
+    // AND write the json context
+    writeln!(file, "{input}").expect("Couldn't write to file");
+
+    // WHEN
+    let params = Parameters::import(file_path.to_str().unwrap()).expect("Error raised");
+
+    // THEN
+    assert_json_snapshot!("parameters_import", params)
 }
